# account-service/docs/openapi.yml
openapi: 3.1.0
info:
  title: Account service API
  description: API for accounts
  version: "1.0"
servers:
- url: http://localhost:8084
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /client/payments:
    post:
      tags:
      - client-controller
      summary: Create a new payment
      operationId: newPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPaymentDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
  /client/accounts/{id}/transactions:
    get:
      tags:
      - client-controller
      summary: Get account transactions
      operationId: findAllTransactions
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTransactionResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
components:
  schemas:
...
    NewPaymentDto:
      type: object
      properties:
        nazivPrimaoca:
          type: string
          minLength: 1
        racunPrimaoca:
          type: string
          minLength: 1
          pattern: "\\d{19}"
        iznos:
          type: number
        pozivNaBroj:
          type: string
        sifraPlacanja:
          type: string
          pattern: "\\d{3}"
        svrhaPlacanja:
          type: string
          minLength: 1
        racunPratioca:
          type: string
          minLength: 1
          pattern: "\\d{19}"
      required:
      - iznos
      - nazivPrimaoca
      - racunPratioca
      - racunPrimaoca
      - sifraPlacanja
      - svrhaPlacanja
...
    ErrorResponseDto:
      type: object
      properties:
        errorCode:
          type: string
        errorTitle:
          type: string
        errorDesc:
          type: string
        timestamp:
          type: string
          format: date-time
        validationErrors:
          type: object
          additionalProperties:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageTransactionResponseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/TransactionResponseDto"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    TransactionResponseDto: {}
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT